+++ CHANGELOG.md
@@ -65,6 +65,7 @@
 ### Chore
 
 - [#6588](https://github.com/blockscout/blockscout/pull/6588) - Add latest image build for frontend-main with specific build-args
+- [#6586](https://github.com/blockscout/blockscout/pull/6586) - Migrate Phoenix 1.5 -> 1.6
 - [#6584](https://github.com/blockscout/blockscout/pull/6584) - Vacuum package-lock.json
 - [#6581](https://github.com/blockscout/blockscout/pull/6581) - Dark mode switcher localStorage to cookie in order to support new UI
 - [#6572](https://github.com/blockscout/blockscout/pull/6572) - pending_block_operations table: remove fetch_internal_transactions column